Cleanup + fix for session cached Databases
Did a lot of comment cleanup and fixed some
"documentation" in there.  Additionally added
a default error configuration value for
controllers.  This determines the error that will
be called if no view is returned by entry
controller logic or attached by default.

One small bugfix was done that makes sure a
session cached database is still in the list of
configured databases.

Changed default redirect to 303 (this is most
common).