Made writable directory check non-existent.
This fixes bugs on some platforms that can't handle the loading
recursion for flourish based exceptions.  Exceptions will be
thrown, no doubt, when a user attempts to actually write to it
or in the case of View, for example, when you try to make it
your cache directory.

May or may not add an exception later in the initialization
process.