Fixed how errors get handled.
The registered response is now the default to be
sent.  That is to say, even if we attempt to
send a response object, if another response is
registered it will be sent instead.

The registered response is also not automatically
overridden when registered.  Instead a clear()
method has been added which will return the
currently registered response and set the
registered response back to null.  This way
we can handle trying/catching other controller
actions.  Ex:

try {
    $response = SubController::action();
} catch (Exception $e) {
    $response = Response::clear();
}

We can then inspect the registered response and
handle it how we want.